refactor: extract response models + add SwiftPM test target
Move the wire-format models (`Organization`, `UsageResponse`,
`ExtraUsageResponse`, `ErrorResponse`) and the in-memory `UsageData` /
`ExtraUsageData` storage types out of `ClaudeAPIService.swift` into
`Usage4Claude/Helpers/ClaudeAPIResponseModels.swift`. The locale-aware
formatting methods on those types (`formattedResetsInHours`,
`formattedCompactAmount`, `statusColor`, `currencySymbol`-based output,
etc.) move to `Usage4Claude/Helpers/UsageData+Formatting.swift` as
extensions.

The split lets the response models live free of `L.*`, `Logger`, and
`UserSettings` dependencies — which means a new SwiftPM `Package.swift`
at the repo root can cherry-pick just `ClaudeAPIResponseModels.swift`
into a `Usage4ClaudeCore` library target and run XCTest against it.

Adds 29 tests for the JSON → in-memory transforms:

  Tests/Usage4ClaudeCoreTests/UsageResponseTests.swift (13 tests)
    - 5-hour parses with integer + floating-point utilization
    - 5-hour with missing resets_at still parses percentage
    - 7-day missing field returns 0% placeholder (new — matches
      "show 7-day for new accounts" behavior added in 1192f35)
    - 7-day (utilization=0, resets_at=null) returns 0% placeholder
    - 7-day with real data parses
    - Opus / Sonnet (utilization=0, resets_at=null) → nil sentinel
    - Opus / Sonnet with real data parse
    - resets_at fractional-second rounding (.645 → up, .159 → down)
    - toUsageData() leaves extraUsage nil (separate fetcher)

  Tests/Usage4ClaudeCoreTests/ExtraUsageResponseTests.swift (16 tests)
    - new fields (monthly_limit + Double used_credits) parse correctly
    - is_enabled=false short-circuits regardless of limit
    - monthly_credit_limit fallback when monthly_limit absent
    - legacy fields (spend_limit_amount_cents + balance_cents) fallback
    - new fields take precedence when both present
    - currency uppercased; defaults to "USD" when absent everywhere
    - implicit enabled signal (no is_enabled, limit > 0)
    - used_credits missing defaults to 0
    - cents → dollars (5000 cents = $50.00)
    - fractional cents preserved (1234.5 → $12.345)
    - percentage computed property
    - currencySymbol mapping (USD/EUR/GBP/JPY/CAD/AUD/BRL/INR + fallback)

To run:

    swift test

Behavior of `toUsageData()` and `toExtraUsageData()` is unchanged — this
is a pure refactor + test-coverage addition. Verified main app still
compiles via xcodebuild against the existing Xcode project.
